DAOS-7454 tests: Restoring running network/cart_self_test.py
Avocado was not detecting the network/cart_self_test.py as an
INSTRUMENTED test and as a result was not being executed.  Moving the
definition of the SelfTest class into a new util file resolves the
isssue.

Test-tag: cartselftest
